Added basic animations, and made units clickable.
Tanks now have a sprite, and both have an idle and moving animation.
When the tanks are idle, it will bounce slowly, when it moves, it will
bounce faster.

Clicking a tank will make the camera focus on the clicked tank. It will
follow it until the user presses 'Tab' or pans the camera around the
level. In addition, the camera won't pan if Left mouse click clicked a
unit.